requirements.txt
langchain
langchain-google-genai
reportlab
markdown
streamlit
python-dotenv